+++ CHANGELOG.md
@@ -13,7 +13,7 @@ the code was deployed.
 
 ### Added
 
-- Sentry logs when an unusually high number of stillness alerts occurs in 24 hours at each location (CU-2chw8y0).
+- Sentry logs when an unusually high number of stillness alerts occurs in set interval of time at each location (CU-2chw8y0).
 
 ## [9.3.0] - 2023-04-27